Added a workaround for a rare loop creation (eg3Z.sam).
This happens during bubble merging where 3 paths exist, 2 short ones
and a long one which skips from the end of one short path to the start
of the other short paths.  Merging the 2 short paths causes the long
one to form a loop.  We now just break the long path to remove the
loop, asserting (justified?) that the shortest assembly is the best
assembly.

Improved test data too.