Bug fix to memory allocation in seq2cigar.
Also implemented, but currently disabled, a mechanism for adding in
a pair of consensus sequences to the graph rather than just
reference.  This is an attempt to resolve regions where an entire
allele gets unmapped as no single read spans the region, so there so no
bubble to collapse. (eg 20:5252300)

It's disabled for now though as while it recovers from FN, it also
adds a comparable number of FP as isn't a good trade.