Added experimental, and currently disabled, khash code.
This replaces use of the local hash_table.c, but for correct_errors it
appears to be working slower.

In implementing the khash using code I discovered several memory
inefficiencies, so updated the HashTable code to be non-volatile and
use a string pool for holding keys.  This reduces memory thrashing and
improves speed.

Also discovered the last 3 correct_error calls were infact a NOP due
to min_count of 1 (every key has at least 1 usage!).  I assume this
was changed at some point, but set it back to a min of 2 while I
investigate further.