Improved error correcting.
There was a bug in the average number of hits per kmer (wrong way
around in calc), but concluded that percentile is a better strategy
than average.

We now accept 90% of kmers as valid and error correct to anything in
that part of the distribution.

To do - check quality too.