Added head/tail tip merging to the seq2cigar_new function.
It only does this based on a simple match with no-indels, but this is
probably OK as indels we don't want to accept into the consensus
unless we can verify both sides of it via a longer kmer.

There is a lot of dead code here and elsewhere to cull now.