New -Q min_mqual option.
For reads which are reassembled but do not match the assembly (and
once were aligned), we now have the option of keeping their alignment
to their original location if and only if they have a minimum mapping
quality and a trivial CIGAR string (xM).

Keeping them in other cases seems to harm false positives too much,
but this looks to be a good middle ground to recover clusters of
neighbouring SNPs.